[Font Lock] Heuristic for finding types
Assume that identifiers that start off capitalised are "types",
similar to the technique used in `rust-mode`.

Note this also gets applied to constructs like:

```
const EMyError: u8 = 1;
```

Which is not a type, but is similar to an enumerand in Rust (which is
highlighted as a type in `rust-mode` as well).