[Refill] Fix continuation lines
Continuation lines are detected by looking for:

- An assignment character (=)
- That isn't inside a comment
- Bounded by the most deeply nested scope

But this means that if it sees an equality or inequality operator, it
will also treat that like an assignment.

The fix is to additionally check that the previous character to the
`=` is not a character that would make it predicate operator token.